+++ src/spotify_playlist_exporter_v2/__init__.py
@@ -1,3 +1,3 @@
 """Spotify Playlist Exporter V2 package."""
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"